Added data storage/retrieval methods
Added four methods for IControllers; getProfileNames, getProfile,
saveProfile and deleteProfile. Use getProfileNames to display names,
then once user has picked a name send the string to getProfile to
retrieve profile data. Also modified one or two small things e.g.
numGamesPlayed now displays without decimal point.

+++ src/ProfilePanel.java
@@ -39,7 +39,7 @@ public class ProfilePanel extends JPanel {
 		
 		//All the stats related to 
 		gamesPlayed = new JLabel();
-		gamesPlayed.setText("" + p.getNumGamesPlayed());
+		gamesPlayed.setText("" + (int) p.getNumGamesPlayed());
 		gamesPlayed.setAlignmentX(Component.CENTER_ALIGNMENT);
 		wlRatioH = new JLabel();
 		wlRatioH.setText("" + p.getWLRatioH());
